+++ src/Mpociot/VatCalculator/VatCalculator.php
@@ -688,22 +688,23 @@ class VatCalculator
         $countryCode = substr($vatNumber, 0, 2);
         $vatNumber = substr($vatNumber, 2);
 
-        if (strtoupper($countryCode) === 'GB' && extension_loaded('curl')) {
-            $curl_handle = curl_init();
-            curl_setopt($curl_handle, CURLOPT_URL, "$this->ukValidationEndpoint/organisations/vat/check-vat-number/lookup/$vatNumber");
-            curl_setopt($curl_handle, CURLOPT_CONNECTTIMEOUT, 2);
-            curl_setopt($curl_handle, CURLOPT_RETURNTRANSFER, 1);
-            curl_setopt($curl_handle, CURLOPT_USERAGENT, 'VatCalculator');
-            $result = curl_exec($curl_handle);
-            curl_close($curl_handle);
+        if (strtoupper($countryCode) === 'GB') {
+            $apiHeaders = get_headers("$this->ukValidationEndpoint/organisations/vat/check-vat-number/lookup/$vatNumber");
+            $apiHeaders = explode(' ', $apiHeaders[0]);
+            $apiStatusCode = (int) $apiHeaders[1];
 
-            $response = json_decode($result, true, 512, JSON_OBJECT_AS_ARRAY);
-
-            if (isset($response['code'])) {
+            if ($apiStatusCode === 400 || $apiStatusCode === 404) {
                 return false;
             }
 
-            return $response['target'];
+            if ($apiStatusCode === 200) {
+                $apiResponse = file_get_contents("$this->ukValidationEndpoint/organisations/vat/check-vat-number/lookup/$vatNumber");
+                $apiResponse = json_decode($apiResponse, true);
+
+                return $apiResponse['target'];
+            }
+
+            throw new VATCheckUnavailableException("The UK VAT check service is currently unavailable (status code $apiStatusCode). Please try again later.");
         } else {
             $this->initSoapClient();
             $client = $this->soapClient;
@@ -722,6 +723,7 @@ class VatCalculator
                     return false;
                 }
             }
+
             throw new VATCheckUnavailableException('The VAT check service is currently unavailable. Please try again later.');
         }
     }